Revert "Try starting a different test server for every test"
This reverts commit 53153ee36bd1f16a6ef5773232b2a0db8f2175c7.
